backend/requirements.txt
# Flask and core dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.23
supabase==2.3.0

# Vector database and embeddings
qdrant-client==1.12.1
sentence-transformers==2.2.2

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Clerk authentication
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0

# Utilities
werkzeug==3.0.1
